% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\title{Generalized Uniform Partition Distribution}
\usage{
experiment(n_items, n_clusters, n_samples, concentration)
}
\arguments{
\item{n_items}{An integer for the number of items in the partition.}

\item{n_clusters}{An integer for the number of clusters in the partition.}

\item{n_samples}{An integer for the number of sampled partitions.}

\item{a}{Parameter}
}
\value{
A function to evaluate the Generalized Uniform Partition Distribution.
}
\description{
A function implementing the probability mass function is provided based on
the supplied probabilty vector and number of items.
}
\examples{
experiment(10, 3, 5, 1)

}
